Adding workday loader
Description: Adding workday loader and checkin loader
Users specify an predefined interval(weekly, monthly) to make the query for workdays.
An List of Workday is show to the user and clicking on an item shows the checkins for the specific workday.

We still have to define and implement the layout for both lists

Commit: Add

TaskNumber: #9